port many datetime64 tests from class-organized tests
