+++ pandas/tests/series/test_arithmetic.py
@@ -79,132 +79,6 @@ class TestSeriesComparison(object):
             assert result.name == names[2]
 
 
-class TestTimestampSeriesComparison(object):
-    def test_dt64_ser_cmp_date_warning(self):
-        # https://github.com/pandas-dev/pandas/issues/21359
-        # Remove this test and enble invalid test below
-        ser = pd.Series(pd.date_range('20010101', periods=10), name='dates')
-        date = ser.iloc[0].to_pydatetime().date()
-
-        with tm.assert_produces_warning(FutureWarning) as m:
-            result = ser == date
-        expected = pd.Series([True] + [False] * 9, name='dates')
-        tm.assert_series_equal(result, expected)
-        assert "Comparing Series of datetimes " in str(m[0].message)
-        assert "will not compare equal" in str(m[0].message)
-
-        with tm.assert_produces_warning(FutureWarning) as m:
-            result = ser != date
-        tm.assert_series_equal(result, ~expected)
-        assert "will not compare equal" in str(m[0].message)
-
-        with tm.assert_produces_warning(FutureWarning) as m:
-            result = ser <= date
-        tm.assert_series_equal(result, expected)
-        assert "a TypeError will be raised" in str(m[0].message)
-
-        with tm.assert_produces_warning(FutureWarning) as m:
-            result = ser < date
-        tm.assert_series_equal(result, pd.Series([False] * 10, name='dates'))
-        assert "a TypeError will be raised" in str(m[0].message)
-
-        with tm.assert_produces_warning(FutureWarning) as m:
-            result = ser >= date
-        tm.assert_series_equal(result, pd.Series([True] * 10, name='dates'))
-        assert "a TypeError will be raised" in str(m[0].message)
-
-        with tm.assert_produces_warning(FutureWarning) as m:
-            result = ser > date
-        tm.assert_series_equal(result, pd.Series([False] + [True] * 9,
-                                                 name='dates'))
-        assert "a TypeError will be raised" in str(m[0].message)
-
-    @pytest.mark.skip(reason="GH-21359")
-    def test_dt64ser_cmp_date_invalid(self):
-        # GH#19800 datetime.date comparison raises to
-        # match DatetimeIndex/Timestamp.  This also matches the behavior
-        # of stdlib datetime.datetime
-        ser = pd.Series(pd.date_range('20010101', periods=10), name='dates')
-        date = ser.iloc[0].to_pydatetime().date()
-        assert not (ser == date).any()
-        assert (ser != date).all()
-        with pytest.raises(TypeError):
-            ser > date
-        with pytest.raises(TypeError):
-            ser < date
-        with pytest.raises(TypeError):
-            ser >= date
-        with pytest.raises(TypeError):
-            ser <= date
-
-    def test_dt64ser_cmp_period_scalar(self):
-        ser = Series(pd.period_range('2000-01-01', periods=10, freq='D'))
-        val = Period('2000-01-04', freq='D')
-        result = ser > val
-        expected = Series([x > val for x in ser])
-        tm.assert_series_equal(result, expected)
-
-        val = ser[5]
-        result = ser > val
-        expected = Series([x > val for x in ser])
-        tm.assert_series_equal(result, expected)
-
-    @pytest.mark.parametrize("left,right", [
-        ("lt", "gt"),
-        ("le", "ge"),
-        ("eq", "eq"),
-        ("ne", "ne"),
-    ])
-    def test_timestamp_compare_series(self, left, right):
-        # see gh-4982
-        # Make sure we can compare Timestamps on the right AND left hand side.
-        ser = pd.Series(pd.date_range("20010101", periods=10), name="dates")
-        s_nat = ser.copy(deep=True)
-
-        ser[0] = pd.Timestamp("nat")
-        ser[3] = pd.Timestamp("nat")
-
-        left_f = getattr(operator, left)
-        right_f = getattr(operator, right)
-
-        # No NaT
-        expected = left_f(ser, pd.Timestamp("20010109"))
-        result = right_f(pd.Timestamp("20010109"), ser)
-        tm.assert_series_equal(result, expected)
-
-        # NaT
-        expected = left_f(ser, pd.Timestamp("nat"))
-        result = right_f(pd.Timestamp("nat"), ser)
-        tm.assert_series_equal(result, expected)
-
-        # Compare to Timestamp with series containing NaT
-        expected = left_f(s_nat, pd.Timestamp("20010109"))
-        result = right_f(pd.Timestamp("20010109"), s_nat)
-        tm.assert_series_equal(result, expected)
-
-        # Compare to NaT with series containing NaT
-        expected = left_f(s_nat, pd.Timestamp("nat"))
-        result = right_f(pd.Timestamp("nat"), s_nat)
-        tm.assert_series_equal(result, expected)
-
-    def test_timestamp_equality(self):
-        # GH#11034
-        ser = pd.Series([pd.Timestamp('2000-01-29 01:59:00'), 'NaT'])
-        result = ser != ser
-        tm.assert_series_equal(result, pd.Series([False, True]))
-        result = ser != ser[0]
-        tm.assert_series_equal(result, pd.Series([False, True]))
-        result = ser != ser[1]
-        tm.assert_series_equal(result, pd.Series([True, True]))
-
-        result = ser == ser
-        tm.assert_series_equal(result, pd.Series([True, False]))
-        result = ser == ser[0]
-        tm.assert_series_equal(result, pd.Series([True, False]))
-        result = ser == ser[1]
-        tm.assert_series_equal(result, pd.Series([False, False]))
-
-
 class TestTimedeltaSeriesComparisons(object):
     def test_compare_timedelta_series(self):
         # regresssion test for GH5963
